Refactor post-processing unit-test to reuse the configuration class by a 3rd party
