+++ unit-tests/unit-tests-post-processing.h
@@ -1,0 +1,298 @@
+// License: Apache 2.0. See LICENSE file in root directory.
+// Copyright(c) 2018 Intel Corporation. All Rights Reserved.
+
+// Definitions and utility functions to load and parse prerecorded frame 
+// utilized in post-processing filters validation
+// The file is intended to be used by both librealsense and 3rd party tools
+#include <vector>
+#include <iostream>
+#include <fstream>
+#include <string>
+#include <map>
+
+struct ppf_test_config
+{
+    std::string name = "";
+    bool        spatial_filter = false;
+    float       spatial_alpha = 0.f;
+    uint8_t     spatial_delta = 0;
+    int         spatial_iterations;
+    bool        temporal_filter = false;
+    float       temporal_alpha = 0.f;
+    uint8_t     temporal_delta = 0;
+    uint8_t     temporal_persistence = 0;
+    int         holes_filling_mode = 0;
+    int         downsample_scale = 1;
+    float       depth_units = 0.001f;
+    float       stereo_baseline = 0.f;
+    float       focal_length = 0.f;
+    uint32_t    input_res_x = 0;
+    uint32_t    input_res_y = 0;
+    uint32_t    output_res_x = 0;
+    uint32_t    output_res_y = 0;
+    std::vector<uint8_t> _input_data;   // stores the actul pixel values
+    std::vector<uint8_t> _output_data;
+
+    void reset()
+    { 
+        name.clear();
+        spatial_filter = temporal_filter = false;
+        spatial_alpha = temporal_alpha = 0;
+        spatial_iterations = temporal_persistence = holes_filling_mode = spatial_delta = temporal_delta =0;
+        input_res_x = input_res_y = output_res_x = output_res_y = 0;
+        downsample_scale = 1;
+        _input_data.clear();
+        _output_data.clear();
+    }
+};
+
+struct synthetic_frame
+{
+    uint16_t res_x, res_y, bpp;
+    std::vector<uint8_t> frame;
+};
+
+std::vector<uint8_t> load_from_binary(const std::string& str)
+{
+    // open the file:
+    std::basic_ifstream<uint8_t> file(str.c_str(), std::ios::binary);
+
+    // read binary data:
+    return std::vector<uint8_t>((std::istreambuf_iterator<uint8_t>(file)), std::istreambuf_iterator<uint8_t>());
+}
+
+enum metadata_attrib : uint8_t
+{
+    res_x,
+    res_y,
+    focal_length,
+    depth_units,
+    stereo_baseline,
+    downscale,
+    spat_filter,
+    spat_alpha,
+    spat_delta,
+    spat_iter,
+    temp_filter,
+    temp_alpha,
+    temp_delta,
+    temp_persist,
+    holes_fill
+};
+
+// The table establishes the mapping of attributes name to be found in the input files
+// generated by the reference viewer (realsense2ex_test.cpp)
+const std::map<metadata_attrib, std::string> metadata_attributes = {
+    { res_x,        "Resolution_x" },
+    { res_y,        "Resolution_y" },
+    { focal_length,  "Focal Length" },
+    { depth_units,   "Depth Units" },
+    { stereo_baseline,"Stereo Baseline" },
+    { downscale,    "Scale" },
+    { spat_filter,   "Spatial Filter Params:" },
+    { spat_alpha,   "SpatialAlpha" },
+    { spat_delta,   "SpatialDelta" },
+    { spat_iter,    "SpatialIterations" },
+    { temp_filter,  "Temporal Filter Params:" },
+    { temp_alpha,   "TemporalAlpha" },
+    { temp_delta,   "TemporalDelta" },
+    { temp_persist, "TemporalPersistency" },
+    { holes_fill,   "HolesFilling" }
+};
+
+using csv_kvp = std::map<std::string, std::string>;
+
+// Parse frame's metadata files and partially fill the configuration struct
+inline ppf_test_config attrib_from_csv(const std::string& str)
+{
+    // First, fill all key/value pair into and appropriate container
+    csv_kvp dict;
+    std::ifstream  data(str.c_str());
+    std::string line;
+    int invalid_line = 0;
+
+    while (true)
+    {
+        std::string key, value;
+        
+        if (!std::getline(data, key, ','))
+            invalid_line++;
+        
+        std::getline(data, value, '\n');
+        dict[key] = value;
+        std::cout << key << ":" << value << std::endl;
+
+        if (invalid_line > 1)  // Two or more non-kvp lines designate eof. Note this when creating the attributes
+            break;
+    }
+
+    // Then parse the data according to the predefined set of attribute names
+    ppf_test_config cfg;
+
+    // Note that the function does not separate between input and output frame metadata.
+    // It is up to the user to retrieve and use the parameters appropriately
+    cfg.input_res_x = dict.count(metadata_attributes.at(res_x)) ? std::stoi(dict.at(metadata_attributes.at(res_x))) : 0;
+    cfg.input_res_y = dict.count(metadata_attributes.at(res_y)) ? std::stoi(dict.at(metadata_attributes.at(res_y))) : 0;
+    cfg.stereo_baseline = dict.count(metadata_attributes.at(stereo_baseline)) ? std::stof(dict.at(metadata_attributes.at(stereo_baseline))) : 0.f;
+    cfg.depth_units = dict.count(metadata_attributes.at(depth_units)) ? std::stof(dict.at(metadata_attributes.at(depth_units))) : 0.f;
+    cfg.focal_length = dict.count(metadata_attributes.at(focal_length)) ? std::stof(dict.at(metadata_attributes.at(focal_length))) : 0.f;
+
+    cfg.downsample_scale = dict.count(metadata_attributes.at(downscale)) ? std::stoi(dict.at(metadata_attributes.at(downscale))) : 1;
+    cfg.spatial_filter = dict.count(metadata_attributes.at(spat_filter)) > 0;
+    cfg.spatial_alpha = dict.count(metadata_attributes.at(spat_alpha)) ? std::stof(dict.at(metadata_attributes.at(spat_alpha))) : 0.f;
+    cfg.spatial_delta = dict.count(metadata_attributes.at(spat_delta)) ? std::stoi(dict.at(metadata_attributes.at(spat_delta))) : 0;
+    cfg.spatial_iterations = dict.count(metadata_attributes.at(spat_iter)) ? std::stoi(dict.at(metadata_attributes.at(spat_iter))) : 0;
+    cfg.temporal_filter = dict.count(metadata_attributes.at(temp_filter)) > 0;
+    cfg.temporal_alpha = dict.count(metadata_attributes.at(temp_alpha)) ? std::stof(dict.at(metadata_attributes.at(temp_alpha))) : 0.f;
+    cfg.temporal_delta = dict.count(metadata_attributes.at(temp_delta)) ? std::stoi(dict.at(metadata_attributes.at(temp_delta))) : 0;
+    cfg.temporal_persistence = dict.count(metadata_attributes.at(temp_persist)) ? std::stoi(dict.at(metadata_attributes.at(temp_persist))) : 0;
+
+    return cfg;
+}
+
+inline bool load_test_configuration(const std::string test_name, ppf_test_config& test_config)
+{
+    bool res = true;
+    std::string base_name = get_folder_path(special_folder::temp_folder) + test_name;
+    static const std::map<std::string, std::string> test_file_names = {
+        { "Input_pixels",    ".Input.raw" },
+        { "Input_metadata",  ".Input.csv" },
+        { "Output_pixels",   ".Output.raw"},
+        { "Output_metadata", ".Output.csv"}
+    };
+    
+    // Verify that all the required test files are present
+    for (auto& filename: test_file_names)
+    { 
+        CAPTURE(base_name);
+        CAPTURE(filename.second);
+        REQUIRE(file_exists(base_name + filename.second));
+    }
+    
+    // Prepare the configuration data set
+    test_config.reset();
+
+    test_config.name = test_name;
+    ppf_test_config input_meta_params = attrib_from_csv(base_name + test_file_names.at("Input_metadata"));
+    ppf_test_config output_meta_params = attrib_from_csv(base_name + test_file_names.at("Output_metadata"));
+    test_config._input_data = std::move(load_from_binary(base_name + test_file_names.at("Input_pixels")));
+    test_config._output_data = std::move(load_from_binary(base_name + test_file_names.at("Output_pixels")));
+    test_config.input_res_x = input_meta_params.input_res_x;
+    test_config.input_res_y = input_meta_params.input_res_y;
+    test_config.output_res_x = output_meta_params.input_res_x;
+    test_config.output_res_y = output_meta_params.input_res_y;
+    test_config.depth_units = input_meta_params.depth_units;
+    test_config.focal_length = input_meta_params.focal_length;
+    test_config.stereo_baseline = input_meta_params.stereo_baseline;
+    test_config.downsample_scale = output_meta_params.downsample_scale;
+    test_config.spatial_filter = output_meta_params.spatial_filter;
+    test_config.spatial_alpha = output_meta_params.spatial_alpha;
+    test_config.spatial_delta = output_meta_params.spatial_delta;
+    test_config.spatial_iterations = output_meta_params.spatial_iterations;
+    test_config.holes_filling_mode = output_meta_params.holes_filling_mode;
+    test_config.temporal_filter = output_meta_params.temporal_filter;
+    test_config.temporal_alpha = output_meta_params.temporal_alpha;
+    test_config.temporal_delta = output_meta_params.temporal_delta;
+    test_config.temporal_persistence = output_meta_params.temporal_persistence;
+    
+    // Perform sanity tests on the input data
+    CAPTURE(test_config.name);
+    CAPTURE(test_config.input_res_x);
+    CAPTURE(test_config.input_res_y);
+    CAPTURE(test_config._input_data.size());
+    CAPTURE(test_config.output_res_x);
+    CAPTURE(test_config.output_res_y);
+    CAPTURE(test_config._output_data.size());
+    CAPTURE(test_config.spatial_filter);
+    CAPTURE(test_config.spatial_alpha);
+    CAPTURE(test_config.spatial_delta);
+    CAPTURE(test_config.spatial_iterations);
+    CAPTURE(test_config.temporal_filter);
+    CAPTURE(test_config.temporal_alpha);
+    CAPTURE(test_config.temporal_delta);
+    CAPTURE(test_config.temporal_persistence);
+    CAPTURE(test_config.holes_filling_mode);
+    CAPTURE(test_config.downsample_scale);
+
+
+    // Basic Sanity
+    REQUIRE(test_config.output_res_x == test_config.input_res_x / test_config.downsample_scale); // Need to be adjusted for rounding up scale factors
+    REQUIRE(test_config.output_res_y == test_config.input_res_y / test_config.downsample_scale); // Need to be adjusted for rounding up scale factors
+    REQUIRE(test_config.input_res_x > 0);
+    REQUIRE(test_config.input_res_y > 0);
+    REQUIRE(test_config.output_res_x > 0);
+    REQUIRE(test_config.output_res_y > 0);
+    REQUIRE(std::fabs(test_config.stereo_baseline) > 0.f);
+    REQUIRE(test_config.depth_units > 0);
+    REQUIRE(test_config.focal_length > 0);
+    REQUIRE((test_config.input_res_x * test_config.input_res_y * 2 )== test_config._input_data.size()); // Assuming uint16_t type
+    REQUIRE((test_config.output_res_x * test_config.output_res_y * 2) == test_config._output_data.size());
+
+    // More specifc test that are awaer of intrinsic of the filters implementation
+    //Note that the specific parameter threshold are correct as of March 2018.
+    if (test_config.spatial_filter)
+    {
+        REQUIRE(test_config.spatial_alpha >= 0.25f);
+        REQUIRE(test_config.spatial_alpha <= 1.f);
+        REQUIRE(test_config.spatial_delta >= 1);
+        REQUIRE(test_config.spatial_delta <= 50);
+    }
+    if (test_config.temporal_filter)
+    {
+        REQUIRE(test_config.spatial_iterations >= 1);
+        REQUIRE(test_config.spatial_iterations <= 5);
+        REQUIRE(test_config.temporal_alpha >= 0.f);
+        REQUIRE(test_config.temporal_alpha <= 1.f);
+        REQUIRE(test_config.temporal_delta >= 1);
+        REQUIRE(test_config.temporal_delta <= 100);
+        REQUIRE(test_config.temporal_persistence >= 0);
+        REQUIRE(test_config.temporal_persistence <= 8);
+    }
+    
+    //TODO: holes_filling mode verification
+    
+    return res;
+}
+
+template <typename T>
+inline bool profile_diffs(const std::string& plot_name, std::vector<T>& distances, const float max_allowed_std, const float outlier)
+{
+    static_assert((std::is_arithmetic<T>::value), "Profiling is defined for built-in arithmetic types");
+
+    std::ofstream output_file(plot_name);
+    for (const auto &val : distances) output_file << val << "\n";
+    output_file.close();
+
+    REQUIRE(!distances.empty());
+
+    float mean = std::accumulate(distances.begin(),
+        distances.end(), 0.0f) / distances.size();
+
+    auto min = std::min_element(distances.begin(), distances.end());
+    auto max = std::max_element(distances.begin(), distances.end());
+    float min_val = *min;
+    float max_val = *max;
+    float e = 0;
+    float inverse = 1.f / distances.size();
+    for (auto elem : distances)
+    {
+        e += pow(elem - mean, 2);
+    }
+
+    float standard_deviation = static_cast<float>(sqrt(inverse * e));
+
+    CAPTURE(mean);
+    CAPTURE(min_val);
+    CAPTURE(max_val);
+    CAPTURE(outlier);
+    CAPTURE(standard_deviation);
+    CAPTURE(max_allowed_std);
+
+    REQUIRE(standard_deviation < max_allowed_std);
+    REQUIRE(fabs((min_val)) < outlier);
+    REQUIRE(fabs((max_val)) < outlier);
+
+    return (fabs(min_val) < outlier) &&
+        (fabs(max_val) < outlier) &&
+        (standard_deviation < max_allowed_std);
+}
